+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [1.0.4] - 2021-02-21
 ### Added
+- new plugin options `useAlias` , `AliasPrefix` .
+- optimize build time when using aliases by fitlering alias from regular imports using a prefix.
+### changed
+- using aliases is not by default now you have to enable it through plugin options.
+
+
+## [1.0.4] - 2021-02-21
+### Added
 - Add the Support to alias paths
-
-